fix: add canonical URLs and correct site host metadata
