+++ app/layout.tsx
@@ -14,7 +14,7 @@ const ibmPlexMono = IBM_Plex_Mono({
 })
 
 export const metadata: Metadata = {
-  metadataBase: new URL('https://slopus.github.io'),
+  metadataBase: new URL('https://happy.engineering'),
   title: 'Happy - Claude Code Mobile Client',
   description: 'Free, open-source mobile app for Claude Code. Control Claude AI from your phone with end-to-end encryption and seamless workflow. Get started with npm install -g happy-coder.',
   keywords: ['Claude', 'AI', 'mobile', 'coding', 'voice-to-code', 'open source', 'npm', 'development', 'programming'],
@@ -25,7 +25,7 @@ export const metadata: Metadata = {
   openGraph: {
     title: 'Happy - Claude Code Mobile Client',
     description: 'Free, open-source mobile app for Claude Code. Control Claude AI from your phone with end-to-end encryption and seamless workflow.',
-    url: 'https://slopus.github.io',
+    url: 'https://happy.engineering/',
     siteName: 'Happy',
     images: [
       {
